+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 ### Changed
 ### Fixed
+- [Issue 123](https://github.com/aza547/wow-recorder/issues/123) - Snap to resolutions within 2 pixels (up from 1).
+- [Issue 128](https://github.com/aza547/wow-recorder/issues/128) - Guard against multiple buffer restarts.
 
 ## [2.7.0] - 2022-09-19
 ### Added